rename parent and child plot to plot and plot list
