+++ groovy-webapp/src/app/components/plot-list/plot-list.component.html
