+++ groovy-webapp/src/app/components/plot/plot.component.html
