+++ groovy-webapp/src/app/components/story/story.component.html
@@ -7,43 +7,43 @@
 
 
 <div class="container" >
-  <label>{{story.id}}</label>
-  <form ngForm>
+  <label>{{story.id}}</label>  <form ngForm>
+  <div class="input-group mb-3">
+    <div class="input-group-prepend">
+      <span class="input-group-text" id="new-title">Title:</span>
+    </div>
+    <input class="form-control col-md-8" type="text" name="title" id="es-title" [(ngModel)]="story.title"/>
+  </div>
+
+  <div class="input-group mb-3">
+    <div class="input-group-prepend">
+      <span class="input-group-text" id="new-author">Author:</span>
+    </div>
+    <input class="form-control" type="text" name="author" id="es-author"  [(ngModel)]="story.author"/>
+  </div>
+
+  <div class="input-group mb-3">
+    <div class="input-group-prepend">
+      <span class="input-group-text" id="new-genre">Genre:</span>
+    </div>
+    <input class="form-control" type="text" name="genre" id="es-genre" [(ngModel)]="story.genre"/>
+  </div>
+
+  <div class="row">
     <div class="input-group mb-3">
       <div class="input-group-prepend">
-        <span class="input-group-text" id="new-title">Title:</span>
+        <span class="input-group-text" id="new-maguffin">Maguffin:</span>
       </div>
-      <input class="form-control col-md-8" type="text" name="title" id="es-title" [(ngModel)]="story.title"/>
+      <textarea class="form-control" name="maguffin" id="es-maguffin" [(ngModel)]="story.maguffin"></textarea>
     </div>
 
     <div class="input-group mb-3">
       <div class="input-group-prepend">
-        <span class="input-group-text" id="new-author">Author:</span>
+        <span class="input-group-text" id="new-summary">Summary:</span>
       </div>
-      <input class="form-control" type="text" name="author" id="es-author"  [(ngModel)]="story.author"/>
+      <textarea class="form-control"  name="summary" id="es-summary" [(ngModel)]="story.summary"></textarea>
     </div>
-
-    <div class="input-group mb-3">
-      <div class="input-group-prepend">
-        <span class="input-group-text" id="new-genre">Genre:</span>
-      </div>
-      <input class="form-control" type="text" name="genre" id="es-genre" [(ngModel)]="story.genre"/>
-    </div>
-
-    <div class="row">
-      <div class="input-group mb-3">
-        <div class="input-group-prepend">
-          <span class="input-group-text" id="new-maguffin">Maguffin:</span>
-        </div>
-        <textarea class="form-control" name="maguffin" id="es-maguffin" [(ngModel)]="story.maguffin"></textarea>
-      </div>
-
-      <div class="input-group mb-3">
-        <div class="input-group-prepend">
-          <span class="input-group-text" id="new-summary">Summary:</span>
-        </div>
-        <textarea class="form-control"  name="summary" id="es-summary" [(ngModel)]="story.summary"></textarea>
-      </div>
-    </div>
-  </form>
+  </div>
+</form>
 </div>
+
